Move parsing stuff to parsing namespace
