+++ src/Hive.Versioning/Parsing/ParseHelpers.cs
@@ -6,7 +6,7 @@ using StringPart = System.ReadOnlySpan<char>;
 using StringPart = Hive.Utilities.StringView;
 #endif
 
-namespace Hive.Versioning
+namespace Hive.Versioning.Parsing
 {
     internal static class ParseHelpers
     {
