+++ src/Hive.Versioning/Parsing/RangeParser.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using Hive.Utilities;
 using static Hive.Versioning.VersionRange;
-using static Hive.Versioning.ParseHelpers;
+using static Hive.Versioning.Parsing.ParseHelpers;
 
 #if !NETSTANDARD2_0
 using StringPart = System.ReadOnlySpan<char>;
@@ -10,7 +10,7 @@ using StringPart = System.ReadOnlySpan<char>;
 using StringPart = Hive.Utilities.StringView;
 #endif
 
-namespace Hive.Versioning
+namespace Hive.Versioning.Parsing
 {
 #pragma warning disable IDE0065 // Misplaced using directive
     // Having this inside the namespace makes it *far* shorter
