+++ src/Hive.Versioning/VersionRange.Components.cs
@@ -1,9 +1,9 @@
-﻿using Hive.Versioning.Resources;
-using System;
+﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
+using Hive.Versioning.Resources;
 
 namespace Hive.Versioning
 {
