+++ src/Hive.Versioning/VersionRange.cs
@@ -1,9 +1,10 @@
-﻿using Hive.Utilities;
-using Hive.Versioning.Resources;
-using System;
+﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text;
+using Hive.Utilities;
+using Hive.Versioning.Resources;
+using Hive.Versioning.Parsing;
 
 #if !NETSTANDARD2_0
 using StringPart = System.ReadOnlySpan<char>;
